# src/adapters/db/models/models.py
from sqlalchemy import Column, Integer, String
from sqlalchemy.orm import DeclarativeBase


class Base(DeclarativeBase):
    pass


class ToolModel(Base):
    __tablename__ = "tool"
    id = Column(Integer, primary_key=True, autoincrement=True)
    name = Column(String, unique=True, nullable=False)
    description = Column(String, nullable=True)
    # Optional link to external documentation for the tool
    link = Column(String, nullable=True)